+++ docs/CHANGELOG.rst
@@ -12,6 +12,8 @@ Unreleased
 Changed
 -------
 * Change ``macs14`` helper function to work on unannotated samples
+* **BACKWARD INCOMPATIBLE:** Update cuffquant ``gff`` input to ``annotation`` in
+  helper and test functions
 
 Fixed
 -----